Rename process_full_data to process_raw and process_raw to process_item Improve some typing and imports
